refactor: extract tests to sibling files for 5 more god-files
Mechanical test-sibling extraction. No behavior change.

  memory/ingestion.rs        1891 -> 1677 (-214)
  memory/store/unified/query 1653 -> 1270 (-383)
  memory/ops.rs              1481 -> 1158 (-323)
  config/ops.rs              1362 ->  803 (-559)
  agent/harness/tool_loop.rs 1383 ->  710 (-673)

Same pattern as 92da9ccb / ef411cec.
4863 lib tests pass.
